requirements.txt
qdrant-client
sentence-transformers
fastapi
uvicorn
pymorphy3
tqdm
optimum[onnxruntime]
PyYAML
python-multipart
